it's 2017: migrate codebase to python3
This is mostly print -> print(), removing decode("utf8") and
encode("utf8") calls because internally strings are now already Unicode
(vs. bytes), and changing the shebang to python3 because we're no longer
compatible to Python 2.

In some cases, make cryptic code better readable in the process.

The invented-here version of urlencode is no longer that easy to
implement, so replace it with the respective call to urllib. This
changes the resulting URL quoting to uppercase, so adapt the test
accordingly.

Apply relative imports to tests so running nosetests does not depend on
the current working directory.

Signed-off-by: Roland Hieber <rohieb@rohieb.name>